word-parser-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
lxml==5.3.0
openai==2.26.0
pydantic==2.9.2
python-dotenv==1.0.1
pypinyin==0.53.0
